LibVector: Add LibVector (will be needed for PCI FAST SEARCH)
