+++ libc/include/stdlib.h
@@ -27,7 +27,7 @@ void dtoa(double num, char* buffer, size_t precision);
 void gtoa(double value, char *buffer, int precision);
 void etoa(double value, char* buffer, int precision);
 void* malloc(size_t size);
-void* realloc(void* ptr, size_t old_size, size_t new_size);
+void* realloc(void* ptr, size_t new_size);
 void utoa(unsigned int value, char *buffer, int base);
 void* calloc(size_t num, size_t size);
 unsigned long long strtoull(const char* str, char** endptr, int base);
